pyproject.toml
[project]
name = "clawbot"
version = "0.1.0"
description = "OpenClaw-like local agent that can control Chrome tabs via Playwright and use Dartmouth Chat via langchain_dartmouth."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  # Keep Playwright pinned (and match its required pyee) to avoid resolver conflicts.
  "playwright==1.42.0",
  "pyee==11.0.1",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dartmouth = [
  "langchain-dartmouth>=0.1.0",
]

[project.scripts]
clawbot = "clawbot.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

